refactor(chatgpt): ChatGpt Assistant 삭제 메서드 migration
